Fix breakage when running yas2atom in Perl v5.24+
This was caused by the use of an (apparently) experimental feature which
was removed in Perl v5.24: https://www.perlmonks.org/?node_id=1190788

The fix appears backwards incompatible with older Perls, so we've bumped
the minimum version of Perl required to run this. When on macOS, be sure
to install the latest version of Perl using Homebrew.