Refactor infobox snippets and fix error in 137be0b
`{{subst:CURRENTDATE}}` expands to `{{CURRENTMONTH}} {{CURRENTYEAR}}...`
when the upload form is submitted, instead of the expected "2024-02-22".
This might be because `CURRENTDATE` is actually a template rather than a
"magic word" (a directive handled specially by the Wikitext language).

Also, stop embedding `{{Original upload date}}` inside `{{Odate}}` tags;
the word "original" is repeated twice, which is cluttered and redundant.